backend/requirements.txt
# Core dependencies
fastapi
uvicorn[standard]
pydantic
python-multipart

# Network packet capture
scapy
netifaces

# LLM and AI agents
langgraph
langchain
langchain-openai
langchain-community
openai

# Async support
aiofiles

# Utils
python-dotenv
requests
httpx
